requirements/test.txt
-r global.txt

pytest-django
pytest-playwright